Change validation way
