--- services/validation.js
@@ -1,36 +1,0 @@
-const Joi = require('joi');
-
-const schemaCreateContact = Joi.object({
-  name: Joi.string().min(3).max(30).required(),
-  email: Joi.string().email().required(),
-  phone: Joi.string()
-    .pattern(/^[(][\d]{3}[)]\s[\d]{3}[-][\d]{4}/)
-    .required(),
-});
-
-const schemaUpdateContact = Joi.object({
-  name: Joi.string().min(3).max(30).optional(),
-  email: Joi.string().email().optional(),
-  phone: Joi.string()
-    .pattern(/^[(][\d]{3}[)]\s[\d]{3}[-][\d]{4}/)
-    .optional(),
-});
-
-const validate = (schema, obj, next) => {
-  const { error } = schema.validate(obj);
-  if (error) {
-    return next({
-      status: 400,
-      message: 'Bad request',
-    });
-  }
-  next();
-};
-
-module.exports.createContact = (req, _res, next) => {
-  return validate(schemaCreateContact, req.body, next);
-};
-
-module.exports.updateContact = (req, _res, next) => {
-  return validate(schemaUpdateContact, req.body, next);
-};
